+++ ocean_utils/__init__.py
@@ -1,0 +1,4 @@
+__version__ = '0.3.2'
+
+#  Copyright 2018 Ocean Protocol Foundation
+#  SPDX-License-Identifier: Apache-2.0