+++ ocean_utils/agreements/utils.py
@@ -3,9 +3,8 @@
 #  Copyright 2018 Ocean Protocol Foundation
 #  SPDX-License-Identifier: Apache-2.0
 
-import os
-
 from ocean_utils.agreements.access_sla_template import ACCESS_SLA_TEMPLATE
+from ocean_utils.agreements.compute_sla_template import COMPUTE_SLA_TEMPLATE
 from ocean_utils.agreements.service_types import ServiceTypes
 
 
@@ -17,6 +16,8 @@ def get_sla_template(service_type=ServiceTypes.ASSET_ACCESS):
     :return: template dict
     """
     if service_type == ServiceTypes.ASSET_ACCESS:
-        return ACCESS_SLA_TEMPLATE.copy()
+        return ACCESS_SLA_TEMPLATE['serviceAgreementTemplate'].copy()
+    elif service_type == ServiceTypes.CLOUD_COMPUTE:
+        return COMPUTE_SLA_TEMPLATE['serviceAgreementTemplate'].copy()
     else:
         raise ValueError(f'Invalid/unsupported service agreement type {service_type}')