+++ ocean_utils/aquarius/aquarius.py
@@ -48,14 +48,13 @@ class Aquarius:
         """Base URL of the aquarius instance."""
         return f'{self._base_url}/ddo'
 
-    def get_service_endpoint(self, did):
+    def get_service_endpoint(self):
         """
         Retrieve the endpoint with the ddo for a given did.
 
-        :param did: Asset DID string
         :return: Return the url of the the ddo location
         """
-        return f'{self.url}/{did}'
+        return f'{self.url}/' + '{did}'
 
     def list_assets(self):
         """
@@ -119,7 +118,7 @@ class Aquarius:
             raise ValueError(response.decode('UTF-8'))
         if parsed_response is None:
             return {}
-        return parsed_response['metadata']
+        return parsed_response['attributes']
 
     def list_assets_ddo(self):
         """
@@ -140,8 +139,8 @@ class Aquarius:
             asset_did = ddo.did
             response = self.requests_session.post(self.url, data=ddo.as_text(),
                                                   headers=self._headers)
-        except AttributeError:
-            raise AttributeError('DDO invalid. Review that all the required parameters are filled.')
+        except AttributeError as e:
+            raise AttributeError(f'DDO invalid. Review that all the required parameters are filled: {e}')
         if response.status_code == 500:
             raise ValueError(
                 f'This Asset ID already exists! \n\tHTTP Error message: \n\t\t{response.text}')