+++ ocean_utils/ddo/public_key_rsa.py
@@ -3,7 +3,6 @@
 
 """
 
-
 #  Copyright 2018 Ocean Protocol Foundation
 #  SPDX-License-Identifier: Apache-2.0