+++ ocean_utils/exceptions.py
@@ -1,4 +1,6 @@
 """Exceptions for ocean_utils """
+
+
 #  Copyright 2018 Ocean Protocol Foundation
 #  SPDX-License-Identifier: Apache-2.0