+++ tests/did_resolver/test_didresolver.py
@@ -5,6 +5,10 @@ import logging
 import secrets
 
 import pytest
+from ocean_keeper import Keeper
+from ocean_keeper.exceptions import (
+    OceanDIDNotFound,
+)
 from web3 import Web3
 
 from ocean_utils.ddo.ddo import DDO
@@ -12,10 +16,6 @@ from ocean_utils.did import DID, did_to_id
 from ocean_utils.did_resolver.did_resolver import (
     DIDResolver,
 )
-from ocean_keeper.exceptions import (
-    OceanDIDNotFound,
-)
-from ocean_keeper import Keeper
 from tests.resources.helper_functions import get_resource_path
 from tests.resources.tiers import e2e_test
 
@@ -37,7 +37,7 @@ def test_did_resolver_library(publisher_account, aquarius):
     sample_ddo_path = get_resource_path('ddo', 'ddo_sample1.json')
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
     asset1 = DDO(json_filename=sample_ddo_path)
-
+    asset1._did = DID.did({"0": "0x1098098"})
     did_registry.register(asset1.asset_id, checksum_test, url=value_test, account=publisher_account)
     aquarius.publish_asset_ddo(asset1)
 
@@ -64,7 +64,7 @@ def test_did_not_found():
 def test_get_resolve_url(aquarius, publisher_account):
     register_account = publisher_account
     did_registry = keeper().did_registry
-    did = DID.did()
+    did = DID.did({"0": "0x1"})
     asset_id = did_to_id(did)
     value_test = aquarius.root_url
     did_resolver = DIDResolver(keeper().did_registry)
@@ -77,16 +77,16 @@ def test_get_resolve_url(aquarius, publisher_account):
 def test_get_resolve_multiple_urls(publisher_account):
     register_account = publisher_account
     did_registry = keeper().did_registry
-    did = DID.did()
-    did2 = DID.did()
-    did3 = DID.did()
-    did4 = DID.did()
-    did5 = DID.did()
-    did6 = DID.did()
-    did7 = DID.did()
-    did8 = DID.did()
-    did9 = DID.did()
-    did10 = DID.did()
+    did = DID.did({"0": "0x1"})
+    did2 = DID.did({"0": "0x2"})
+    did3 = DID.did({"0": "0x3"})
+    did4 = DID.did({"0": "0x4"})
+    did5 = DID.did({"0": "0x5"})
+    did6 = DID.did({"0": "0x6"})
+    did7 = DID.did({"0": "0x7"})
+    did8 = DID.did({"0": "0x8"})
+    did9 = DID.did({"0": "0x9"})
+    did10 = DID.did({"0": "0x10"})
     value_test = 'http://localhost:5000'
     value_test2 = 'http://localhost:5001'
     value_test3 = 'http://localhost:5002'