+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 History
 =======
 
+0.2.5 (October 23rd, 2019)
+-------------------------
+* Fix bug in metadata index
+
+0.2.4 (October 23rd, 2019)
+-------------------------
+* Fix index number in service agreements
+
+0.2.3 (October 22nd, 2019)
+-------------------------
+* Fix compute template
+
+0.2.2 (October 15th, 2019)
+-------------------------
+* Add compute template
+
+0.2.1 (September 12th, 2019)
+---------------------------
+* Add setter for service endpoint.
+
+0.2.0 (September 5th, 2019)
+---------------------------
+* Update methods to changes in the different OEP's 
+  - [OEP7](https://github.com/oceanprotocol/OEPs/tree/master/7/v0.2)
+  - [OEP8](https://github.com/oceanprotocol/OEPs/tree/master/8/v0.4)
+  - [OEP11](https://github.com/oceanprotocol/OEPs/tree/master/11/v0.2)
+  - [OEP12](https://github.com/oceanprotocol/OEPs/tree/master/12)
+  - [OEP13](https://github.com/oceanprotocol/OEPs/tree/master/13)
+
 0.1.0 (Aug, 2019)
 -----------------------
+* First version after split squid-py in different libraries.